+++ stroom-query-testing/src/main/java/stroom/query/testing/FifoLogbackRule.java
@@ -22,6 +22,7 @@ public class FifoLogbackRule extends ExternalResource {
 
     @Override
     protected void after() {
+        FifoLogbackAppender.popLogs();
     }
 
     public void checkAuditLogs(final int expected) {